+++ src/arc/util/serialization/UBJsonReader.java
@@ -65,15 +65,15 @@ public class UBJsonReader implements BaseJsonReader{
         else if(type == 'F')
             return new JsonValue(false);
         else if(type == 'B')
-            return new JsonValue((long)readUChar(din));
+            return new JsonValue(readUChar(din));
         else if(type == 'U')
-            return new JsonValue((long)readUChar(din));
+            return new JsonValue(readUChar(din));
         else if(type == 'i')
             return new JsonValue(oldFormat ? (long)din.readShort() : (long)din.readByte());
         else if(type == 'I')
             return new JsonValue(oldFormat ? (long)din.readInt() : (long)din.readShort());
         else if(type == 'l')
-            return new JsonValue((long)din.readInt());
+            return new JsonValue(din.readInt());
         else if(type == 'L')
             return new JsonValue(din.readLong());
         else if(type == 'd')
@@ -193,7 +193,7 @@ public class UBJsonReader implements BaseJsonReader{
         if(type == 'S'){
             size = parseSize(din, true, -1);
         }else if(type == 's')
-            size = (long)readUChar(din);
+            size = readUChar(din);
         else if(sOptional) size = parseSize(din, type, false, -1);
         if(size < 0) throw new ArcRuntimeException("Unrecognized data type, string expected");
         return size > 0 ? readString(din, size) : "";
@@ -205,15 +205,15 @@ public class UBJsonReader implements BaseJsonReader{
 
     protected long parseSize(final DataInputStream din, final byte type, final boolean useIntOnError, final long defaultValue)
     throws IOException{
-        if(type == 'i') return (long)readUChar(din);
-        if(type == 'I') return (long)readUShort(din);
+        if(type == 'i') return readUChar(din);
+        if(type == 'I') return readUShort(din);
         if(type == 'l') return readUInt(din);
         if(type == 'L') return din.readLong();
         if(useIntOnError){
             long result = (long)((short)type & 0xFF) << 24;
             result |= (long)((short)din.readByte() & 0xFF) << 16;
             result |= (long)((short)din.readByte() & 0xFF) << 8;
-            result |= (long)((short)din.readByte() & 0xFF);
+            result |= (short)din.readByte() & 0xFF;
             return result;
         }
         return defaultValue;
@@ -228,7 +228,7 @@ public class UBJsonReader implements BaseJsonReader{
     }
 
     protected long readUInt(final DataInputStream din) throws IOException{
-        return ((long)din.readInt());
+        return din.readInt();
     }
 
     protected String readString(final DataInputStream din, final long size) throws IOException{
